remove_empties option to removes empty strings
A common cleaning task is removing columns and or rows that are composed uniquely of length 0 "void" strings, i.e. strings which only entries are ''. They are not detect by `is.na()`. I added the optional argument `drop_empty_string` which if set to true would detect and remove these columns. One liner explanation in the params.